Implement the test framework
Supported laguages are:
- C
- Go
- Python
- Rust
- TypeScript
- Zig

Test scenarios are defined in yaml files.
Some examples are provided in the scenarios directory.
Test grammar is defined in the schema.json file and
described in the documentation.

Add run.sh to wrap the execution of the test framework
for each language.
Call run.sh -f <scenario file> or -d <scenario directory>
to run the test framework.

Reference implementation aka python_cairo can be executed
with:
 -r (for ramdom) to generate random test scenarios.
 --fix to fix the test scenarios.

Signed-off-by: Jean-François <jfm@laposte.net>